Add module query status update to stdout.
Whenever a YoctoPuce module is queried, print that and the modules
friendly name to stdout. This serves as a "I'm alive" beacon (via
"systemctl status yocto_exporter" and apparently as a side effect
solves the problem of the exporter eventually throwing an exception
in the prometheus network code that freezes a single module at its
current value (happened reliably with the light sensor). Why?
No idea, but I suspect Python GIL shenanigans.